files/cloudwatch_agent/requirements.txt
boto
boto3
docopt
mock
PyYAML
tox